requirements.txt
streamlit
statsbombpy
pandas
mplsoccer
matplotlib
seaborn
pillow
time